+++ src/main/java/interface_adapter/chat/promptController.java
@@ -1,6 +1,6 @@
 package interface_adapter.chat;
 
-import use_case.ChatService.PrompInteractor;
+import data_access.gpt_api_calls.PrompInteractor;
 
 import javax.swing.*;
